+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#3385](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3385))
 - `opentelemetry-instrumentation` Make auto instrumentation use the same dependency resolver as manual instrumentation does
   ([#3202](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3202))
+- `opentelemetry-instrumentation-aiokafka` Add instrumentation of `consumer.getmany` (batch)
+  ([#3257](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3257))
 
 ### Fixed
 
@@ -35,8 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Version 1.31.0/0.52b0 (2025-03-12)
 
 ### Added
-- `opentelemetry-instrumentation-aiokafka` Add instrumentation of `consumer.getmany` (batch)
-  ([#3257](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3257))
+
 - `opentelemetry-instrumentation-openai-v2` Update doc for OpenAI Instrumentation to support OpenAI Compatible Platforms
   ([#3279](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3279))
 - `opentelemetry-instrumentation-system-metrics` Add `process` metrics and deprecated `process.runtime` prefixed ones