+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#3385](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3385))
 - `opentelemetry-instrumentation` Make auto instrumentation use the same dependency resolver as manual instrumentation does
   ([#3202](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3202))
+- `opentelemetry-instrumentation-aiokafka` Add instrumentation of `consumer.getmany` (batch)
+  ([#3257](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3257))
 
 ### Fixed